+++ Assets/lilToon/Editor/lilToonSetting.cs
@@ -82,7 +82,6 @@ public class lilToonSetting : ScriptableObject
     public float defaultBeforeExposureLimit = 10000.0f;
     public float defaultMonochromeLighting = 0.0f;
     public float defaultlilDirectionalLightStrength = 1.0f;
-    public Vector4 defaultLightDirectionOverride = new Vector4(0.001f, 0.002f, 0.001f, 0.0f);
 
     public lilToonPreset presetSkin;
     public lilToonPreset presetFace;