+++ Assets/lilToon/CHANGELOG.md
@@ -26,6 +26,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Changed POM to parameter
 - Expanded preset
 - Changed to save `Setup from FBX` presets in shader settings
+- Changed to preserve `Light Direction Override` property in `Fix lighting`
 
 ### Fixed
 - Fixed an issue where fog wasn't working well in URP